# src/main/java/app/jobzy/rest/controller/JobCreationController.java
package app.jobzy.rest.controller;

import app.jobzy.domain.dto.request.CompanyInfoRequestDto;
import app.jobzy.domain.dto.request.JobInfoRequestDto;
import app.jobzy.domain.dto.response.CompanyInfoResponseToken;
import app.jobzy.domain.dto.response.GeneratedVacancyDto;
import app.jobzy.service.CompanyInfoRetrievalService;
import app.jobzy.service.VacancyCreationService;
import jakarta.validation.Valid;
import java.util.UUID;
import lombok.RequiredArgsConstructor;
import lombok.extern.log4j.Log4j2;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.CrossOrigin;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequiredArgsConstructor
@Log4j2
@CrossOrigin(origins = {"http://localhost:3000", "https://dev.jobzy.app"})
public class JobCreationController {

  private final CompanyInfoRetrievalService companyInfoRetrievalService;
  private final VacancyCreationService vacancyCreationService;

  @PostMapping("/create-company-info")
  public ResponseEntity<CompanyInfoResponseToken> createCompanyInfo(
      @RequestBody @Valid CompanyInfoRequestDto companyInfoRequestDto) {
    log.debug("Received companyInfoRequestDto {}", companyInfoRequestDto);
    var response = companyInfoRetrievalService.getCompanyInfoResponseToken(companyInfoRequestDto);
    return new ResponseEntity<>(response, HttpStatus.CREATED);
  }

  @PostMapping("/create-vacancy")
  public ResponseEntity<GeneratedVacancyDto> createVacancyText(
      @RequestBody @Valid JobInfoRequestDto jobInfoRequestDto, @RequestHeader("requestId") UUID requestId) {
    GeneratedVacancyDto generatedVacancy = vacancyCreationService.createVacancy(jobInfoRequestDto, requestId);
    return new ResponseEntity<>(generatedVacancy, HttpStatus.CREATED);
  }
}